Added math expression parsing

+++ FrontEnd/Pages/Views/Elements/ChartItem.cs
@@ -2,10 +2,10 @@ using System;
 
 public class ChartItem
 {
-    public int Value { get; set; }
+    public double Value { get; set; }
     public DateTime Date { get; set; }
 
-    public ChartItem(int value, DateTime date)
+    public ChartItem(double value, DateTime date)
     {
         Value = value;
         Date = date;
